+++ lib/src/clients/http_client.dart
@@ -21,10 +21,49 @@ abstract class HttpClient {
   /// Unlike a standard [Future], a [CancelableOperation] may be cancelled if
   /// the in-flight request is no longer valid or wanted. In that event, the
   /// [CancelableOperation.value] may never complete.
+  ///
+  /// May throw a [HttpClientException] if the request fails.
   CancelableOperation< /*Json*/ Object> requestJson(
     String path, {
     String method: 'GET',
     Map<String, Object> payload,
     Map<String, String> headers,
   });
+}
+
+/// Thrown by [HttpClient] when an HTTP or network error occurs.
+class HttpClientException implements Exception {
+  final int code;
+  final String reason;
+
+  /// An HTTP exception with an error [code] and optional [reason].
+  const HttpClientException(this.code, [this.reason]);
+
+  /// An HTTP exception caused by another [error].
+  const factory HttpClientException.from(
+    Object error,
+  ) = _WrappedHttpClientException;
+
+  /// An HTTP exception caused by unknown causes.
+  const HttpClientException.unknown() : this(null, null);
+
+  @override
+  String toString() {
+    if (code == null) {
+      return 'Unknown';
+    }
+    if (reason == null) {
+      return '$code';
+    }
+    return '$code: $reason';
+  }
+}
+
+class _WrappedHttpClientException extends HttpClientException {
+  final Object _error;
+
+  const _WrappedHttpClientException(this._error) : super(null, null);
+
+  @override
+  String toString() => _error.toString();
 }