+++ lib/src/schema/structures/gateway.g.dart
@@ -74,3 +74,52 @@ class _$GatewayDispatch implements GatewayDispatch {
   @override
   final String name;
 }
+
+class _$GatewayReadyBuilder {
+  int version;
+  User user;
+  List<Channel> privateChannels;
+  String sessionId;
+  List<String> trace;
+  _$GatewayReady _build() {
+    return new _$GatewayReady._internal(
+      version: version,
+      user: user,
+      privateChannels: privateChannels,
+      sessionId: sessionId,
+      trace: trace,
+    );
+  }
+}
+
+class _$GatewayReady implements GatewayReady {
+  const _$GatewayReady._internal(
+      {this.version,
+      this.user,
+      this.privateChannels,
+      this.sessionId,
+      this.trace});
+  factory _$GatewayReady(void build(_$GatewayReadyBuilder builder)) {
+    final builder = new _$GatewayReadyBuilder();
+    build(builder);
+    return builder._build();
+  }
+  factory _$GatewayReady.fromJson(Map<String, Object> json) =>
+      (new _$GatewayReadyBuilder()
+            ..version = json['v'] as int
+            ..user = new User.fromJson(json['user'] as Map<String, Object>)
+            ..privateChannels = json['private_channels'] as List<Channel>
+            ..sessionId = json['session_id'] as String
+            ..trace = json['_trade'] as List<String>)
+          ._build();
+  @override
+  final int version;
+  @override
+  final User user;
+  @override
+  final List<Channel> privateChannels;
+  @override
+  final String sessionId;
+  @override
+  final List<String> trace;
+}